[#49] Refactor: 관련 기능별로 폴더 구조 재정리
