+++ src/routes/payment/paymentRequest.tsx
@@ -1,9 +1,9 @@
-import Header from "../components/headers/Header.tsx";
+import Header from "../../components/headers/Header.tsx";
 import styled from "styled-components";
 import {useLocation, useNavigate} from "react-router-dom";
-import Button from "../components/buttons/Button.tsx";
-import {Wrapper} from "./letterSentConfirm.tsx";
-import BackButton from "../components/buttons/BackButton.tsx";
+import Button from "../../components/buttons/Button.tsx";
+import {Wrapper} from "../letters/letterSentConfirm.tsx";
+import BackButton from "../../components/buttons/BackButton.tsx";
 
 const PaymentWrapper = styled(Wrapper)`
   gap: 50px;
